Slight changes. Not much more functionality.
Only real changes are longitude and latitude changed to floats. Added
segue back to original view controller. For both states and cities for
the information to register you must select the one you want, go back,
then reselect the state/city you want for the information to be
recognized. Not sure how to fix it.